# docs/openapi.yml
openapi: 3.0.3
info:
  title: Stellar GreenPay API
  version: 1.0.0
  description: |
    REST API for the Stellar GreenPay climate donation platform.
    Every donation is recorded on the Stellar blockchain via Soroban smart contracts.
    All `/api/v1/*` JSON responses use a single envelope. Successful responses
    are `{ success: true, data, meta? }`; error responses are
    `{ success: false, error: { code, message, details? } }`. The unversioned
    `/api/*` paths are legacy aliases that redirect to `/api/v1/*`; clients
    should migrate to `/api/v1/*` and branch on `error.code`, never message text.
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development
...
components:
  schemas:
    Success:
      type: object
      required: [success, data]
      additionalProperties: false
      properties:
        success:
          type: boolean
          enum: [true]
          example: true
        data:
          description: Endpoint-specific response payload.
        meta:
          type: object
          additionalProperties: true
          description: Optional response metadata such as pagination or cursors.
    Error:
      type: object
      required: [success, error]
      additionalProperties: false
      properties:
        success:
          type: boolean
          enum: [false]
          example: false
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
              example: PROJECT_NOT_FOUND
            message:
              type: string
              description: Human-readable display message.
              example: Project not found
            details:
              description: Optional structured diagnostic details.
...
    Donation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        donorAddress:
          type: string
        amountXLM:
          type: string
          nullable: true
        amount:
          type: string
        currency:
          type: string
          example: XLM
        message:
          type: string
          nullable: true
        transactionHash:
          type: string
        createdAt:
          type: string
          format: date-time
    Profile:
      type: object
      properties:
        publicKey:
          type: string
        displayName:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        totalDonatedXLM:
          type: string
        projectsSupported:
          type: integer
        badges:
          type: array
          items:
            type: object
            properties:
              tier:
                type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Rating:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        donorAddress:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        review:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
  parameters:
...
    publicKey:
      name: publicKey
      in: path
      required: true
      schema:
        type: string
        example: GABC...XYZ
...
paths:
...
  /api/donations:
    post:
      tags: [Donations]
      summary: Record a donation
      operationId: recordDonation
      description: Records an already-submitted on-chain donation. Idempotent by transactionHash.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, donorAddress, transactionHash]
              properties:
                projectId:
                  type: string
                  format: uuid
                donorAddress:
                  type: string
                  example: GABC...XYZ
                amountXLM:
                  type: number
                  example: 100.0
                amount:
                  type: number
                currency:
                  type: string
                  default: XLM
                message:
                  type: string
                  maxLength: 100
                transactionHash:
                  type: string
                  pattern: '^[a-fA-F0-9]{64}$'
      responses:
        '201':
          description: Donation recorded
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Donation'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
  /api/donations/project/{projectId}:
    get:
      tags: [Donations]
      summary: List donations for a project
      operationId: listProjectDonations
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          description: ISO timestamp for cursor-based pagination
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Paginated donations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Donation'
                      nextCursor:
                        type: string
                        format: date-time
                        nullable: true
  /api/donations/project/{projectId}/messages:
    get:
      tags: [Donations]
      summary: List donations with messages for a project
      operationId: listProjectDonationMessages
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        '200':
          description: Donations with messages, sorted by amount
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Donation'
  /api/donations/donor/{publicKey}:
    get:
      tags: [Donations]
      summary: List donations by a donor
      operationId: listDonorDonations
      parameters:
        - $ref: '#/components/parameters/publicKey'
      responses:
        '200':
          description: All donations by this donor
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Donation'
        '400':
          description: Invalid public key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Profiles ─────────────────────────────────────────────────────────────────
  /api/profiles/{publicKey}:
    get:
      tags: [Profiles]
      summary: Get donor profile
      operationId: getProfile
      parameters:
        - $ref: '#/components/parameters/publicKey'
      responses:
        '200':
          description: Donor profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Profile'
        '400':
          description: Invalid public key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/profiles:
    post:
      tags: [Profiles]
      summary: Create or update donor profile
      operationId: upsertProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [publicKey]
              properties:
                publicKey:
                  type: string
                displayName:
                  type: string
                  maxLength: 30
                bio:
                  type: string
                  maxLength: 300
      responses:
        '200':
          description: Profile created or updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Profile'
        '400':
          description: Invalid public key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Leaderboard ──────────────────────────────────────────────────────────────
...
  /api/ratings:
    post:
      tags: [Ratings]
      summary: Submit or update a project rating
      operationId: submitRating
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, donorAddress, rating]
              properties:
                projectId:
                  type: string
                  format: uuid
                donorAddress:
                  type: string
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                review:
                  type: string
      responses:
        '201':
          description: Rating submitted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Rating'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/ratings/pending:
    get:
      tags: [Ratings]
      summary: Get a project pending a rating from a donor
      operationId: getPendingRating
      description: Returns a project the donor donated to >7 days ago but hasn't rated yet.
      parameters:
        - name: donorAddress
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project pending a rating (or null)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        nullable: true
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
                          name:
                            type: string
        '400':
          description: Missing donorAddress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Admin ────────────────────────────────────────────────────────────────────
...